+++ src/macros.rs
@@ -91,33 +91,3 @@ macro_rules! quick_derive {
         })
     };
 }
-
-#[macro_export]
-macro_rules! derive_trait_with_capacity {
-    ($data:expr, $capacity:expr, _, $($trait:tt)*) => {
-        derive_trait_with_capacity!($data, $capacity, None, _, $($trait)*)
-    };
-    ($data:expr, $capacity:expr, ($($path:tt)*), $($trait:tt)*) => {
-        derive_trait_with_capacity!($data, $capacity, None, ($($path)*), $($trait)*)
-    };
-    ($data:expr, $capacity:expr, $super:expr, _, $($trait:tt)*) => {
-        $crate::__rt::parse2($crate::__rt::quote!($($trait)*))
-            .map_err($crate::Error::from)
-            .and_then(|trait_: $crate::__rt::ItemTrait| {
-                derive_trait_with_capacity!($data, $capacity, $super, $crate::__rt::path(Some(trait_.ident.clone().into())), trait_)
-            })
-    };
-    ($data:expr, $capacity:expr, $super:expr, ($($path:tt)*), $($trait:tt)*) => {
-        $crate::__rt::parse2($crate::__rt::quote!($($path)*))
-            .map_err($crate::Error::from)
-            .and_then(|path| {
-                derive_trait_with_capacity!($data, $capacity, $super, path, $crate::__rt::parse2($crate::__rt::quote!($($trait)*))?)
-            })
-    };
-    ($data:expr, $capacity:expr, $path:expr, $trait:expr) => {
-        derive_trait_with_capacity!($data, $capacity, None, $path, $trait)
-    };
-    ($data:expr, $capacity:expr, $super:expr, $path:expr, $trait:expr) => {
-        $data.impl_trait_with_capacity($capacity, $path, $super, $trait).map($crate::build)
-    };
-}