+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Add quick_derive macro
+* Remove derive_trait_with_capacity macro
 
 # 0.4.0 - 2018-12-22